requirements.txt
python-dotenv==1.0.0
pydantic==1.10.26
SpeechRecognition==3.8.1
PyAudio==0.2.14
gTTS==2.5.0
deep-translator==1.11.4
langchain==0.0.352
langchain_google_genai==0.0.5
google_generativeai==0.3.2
gradio==3.36.1
grpcio==1.60.0
grpcio-tools==1.60.0
gradio_client==0.2.7